Fix field change not clearing error when valid
When a value is set it is automatically passed through validation, and if there is an error it is set on the field.

Setting the same field's value to a valid one and there is already a field error message present, the latter would not get cleared in order to indicate that the field is valid now.