Add shouldValidate option to value changing funcs
This includes:
 - model.setFieldValue()
 - model.handleChange()
 - field.change()

shouldValidate is an optional last parameter to all of these methods, which defaults to `true`. When it is set to `false` the value will not be validated when changed by this call.

This introduces some flexibility to when exactly a value is validated. For example it is possible that we do not want the value to be validated on every keystroke (onChange for example) and instead when the input loses focus (onBlur) or on some other special condition.

---

This commit also improves the jsdoc for these methods and functions and their respecitve types.