Change model's isDirty semantics
It now dynamically checks whether any of the fields is marked as dirty. It
was previously manually set on change and reset, which only introduced
additional complexity into the codebase.

Even though it iterates over all the fields it is highly unlikely that
it would ever turn out to be an issue regarding performance.