Introduce Form component
This is a tiny wrapper around an HTML <form/> element, which sets up its submission and resetting.

This component will rerender itself anytime the model changes, but will not render all of its children thanks to React's rendering semantics in cases where children are the same. This is the best possible solution for optimal performance without forcing users to learn this wrapper trick or extract additional components.

All of the primitives for this component are exposed and it is completely possible to recreate this component in userland and even a modified version of it if necessary.