Change field.dirty semantics
Fields were previously marked as dirty from the first time they were changed onwards. Now they are considered dirty when their current value is different from their initial one. This affects the model's isDirty property in the same way.

Comparison between the new and old value is a simple equality ===, which means that if the initialValue is an object (including arrays), it will always be dirty after its first change.