Add model and field touched state
This is a boolean value set on each field that tells whether a field was changed, validated or otherwise updated since its initialization or the last time it was reset.

The model instance exposes a property isTouched, which tells whether any of the fields is touched.

---

This is different from dirty, which will be false whenever the current field value is the same as its initial one.