Fix isSchemaFieldObject implementation
It used to work in the opposite way of what its name said, meaning that it would return true when the input is NOT a schema field object. It now properly returns true whenever the input is an object containing only keys valid for a schema field object.

This commit also improves related typings, and introduces a test to assert that model can be correctly initialized and built when there are fields with array values.