Rework model.reset() method
reset() method used to reset the whole model instance itself.

Now it has an optional field name parameter. When called with no argument or with an empty string the method does exactly what it used to do prior to this commit.

When called with a field name argument it will find and reset the field with this name. This works the same way that model.resetField() works.

What is new is its ability to reset all nested fields for a partial key. For example calling model.reset("nested") will reset all fields such as nested.foo, nested.foo2, nested.deeper.foo, nested.even.deeper.foo and so on. This allows nested objects or just a part of the model to be reset seamlesssly.