Remove model.resetField()
The decision to remove this method is based on the fact that the model allows multiple fields to be reset at once. This either requires model.reset() or model.resetField() to implement this behaviour or introduce a third, independent method.

Introducing a third method only increases the API surface and adds weight to the bundle.

Implementing multiple field resets for model.resetField() is awkward and does not reflect what its name claims to do. On the other hand model.reset() is a not specific name, which allows it to accept arguments, which subtly tweak its behaviour.

That is why the model.reset() method encapsulates all of this, while sticking to a single purpose -- resetting the whole model or only a part of it.

The API is as follows:
// Reset the entire model.
model.reset()

// Reset a single field.
model.reset("field")

// Reset all nested fields starting with a partial key.
model.reset("nested")