Don't expose model fields property
First of all, this is a major breaking change.

Model object used to expose a fields property, which represented all fields shaped like the target object. In order to support this some complex typings and implementation details are required.

Since it is not recommended to access model fields directly from the model object it seems appropriate to not expose fields as a property directly on the model object. Not only does this simplify the underlying implementation, but it also improves performance, future maintenance and also opens up some new possibilities without all of the burden that would come otherwise.

Individual fields can still be accessed via model.getField() when really necessary. The only downside is that the typescript assitance is lost, but regardless of that an exception will be thrown when a key to invalid field is given. This should be enough of an assitance to not have inconsitent form fields, though a bit more annoying and tedious in some cases.