Add useModel property and method selectors
This adds the ability to call useModel with the string name of a model's property or method and receive it as the result of the hook call. This subscribes the component to changes to the selected property. Does not subscribe to updates when selecting a method since methods never change and is unnecessary to subscribe.

---

This is a way to opt-in for optimal performance when it is necessary to access a model property or method in a component that is heavy to re-render on every single model change.

For example when we would like to use the model submission method directly inside our form, which renders all of the input fields, we would call useModel("submit"). This way when any of the input changes and updates the model, the entire form will not be rerendered, since the submit method has not changed. Only the changed input component would render.

Similarly if we call useModel("isDirty") the calling component will only update when the isDirty value changes.

In contrast if we were to call useModel() and then access model.isDirty, the component will update on every single model change since useModel() subscribes to every single model change.