Add model state
Model state is meant to provide users with a way to associate custom data to the model instance.

For example when a the request to server throws an error during submission it can be caught and model's state can reflect that an error occured. This state can then be used to display an error message to users, deep down in the component tree.

Another useful example is when there is a toggleable optional field. Model's state is the perfect place to save the state of the toggle.