Remove called outside of context exception
When either useField(...) or useModel() with no arguments were called outside of ModelContext an exception would be thrown. This exception is quite unnecessary and sometimes not very clear how to avoid it. Furthermore it was quite long and only added extra bytes to the bundle.

This exception is now removed and instead the ModelContext is initialized with an empty model object. This way useModel() can be called outside of context without throwing and the same goes for useField when called with an empty string argument.

Attempting to access a model field in any way will still throw an exception explaining that the requested field is missing and was not part of the initial schema, which is quite enough of a pointer that context could be missing.