Throw when selecting invalid model property
Calling useModel() with the name of a private or invalid model property used to return 'undefined'. Then it would often result in errors such as 'undefined is not a funcction'.

Instead of such cryptic errors it does not return 'undefined` and instead throws an exception, explaining that an attempt to select an undefined model propertty was made.