Remove useModel schema implementation
It was possible to call useModel with a schema and receive a model object. This seems useful for smaller forms, but is actually not that useful since the model.fields are no longer exposed and model.getField() has to be called to obtain each of the fields.

This implementation can be added back in the future, if nececssary, but currently it unnecessarily increases the API surface and can introduce unwanted complexity and confustion about how the library is meant to be used.

The thinking is that anyone who has to use a library for data modelling would probably need to create more than a single form or at the very least a single, but complex or large one. Regardless of the use case, creating reusable input components which hook onto the model is the recommended pattern.