Fix useModel method selection resulting in crash
Calling a selected model method via useModel(selector) results in a crash. This is due to the fact that it is coppied and has no access to the `this` object, which is undefined for it.

This can be solved by binding each public method explicitly, but a simpler solution is to just transform the public methods into public function properties. The latter close over the `this` object and can use it as intended.