Storage API
This introduces a new opt-in API, which allows users to provide a storage
driver to the model, which will be used to save model data and then load
from it the next time the model instance is recreated. This is essentially
a first class way to remember the user input and have it available on a
subsequent visit.

A number of guards and checks are present in order to make sure that only
saved values which are compatible with the current model schema will be
loaded. This is to prevent crashes due to saved data and schema incompatibility
when a form is patched. While the current implementation is considered to
be sufficient, there are a lot more guards that can be added to make sure
that compatibility is maintained and no crashes will occur.

---

This commit includes only a single WebStorage driver, which as its name
suggests, is meant to be used on the web. It is a simple wrapper around
localStorage.

While it is planned for more drivers, such as NativeStorage, to be introduced,
the API allows for custom drivers to be created at will, so long as they
implement or match the public interface ModelStorage. The model will call
the respective methods when necessary, while the driver is supposed to
handle the saving, loading, clearing and so on.